Make the profile tab work, fix edge cases in profile view
