+++ mobile/views/User/index.js
@@ -1,128 +1,30 @@
 // @flow
 import * as React from 'react';
-import { Text, View, StatusBar, ScrollView } from 'react-native';
 import compose from 'recompose/compose';
-import { getUserById } from '../../../shared/graphql/queries/user/getUser';
-import getUserThreadConnection from '../../../shared/graphql/queries/user/getUserThreadConnection';
-import ViewNetworkHandler from '../../components/ViewNetworkHandler';
-import ThreadFeed from '../../components/ThreadFeed';
-import type { GetUserType } from '../../../shared/graphql/queries/user/getUser';
-
 import {
-  Wrapper,
-  CoverPhoto,
-  CoverPhotoContainer,
-  ProfilePhoto,
-  ProfilePhotoContainer,
-  ProfileDetailsContainer,
-  Name,
-  Username,
-  Description,
-  ThreadFeedTabContainer,
-  ThreadFeedTab,
-  TabLabel,
-} from './style';
+  getUserById,
+  getCurrentUser,
+} from '../../../shared/graphql/queries/user/getUser';
+import ViewNetworkHandler from '../../components/ViewNetworkHandler';
+import Profile from './profile';
 
 type Props = {
-  isLoading: boolean,
-  hasError: boolean,
-  navigation: Object,
-  data: {
-    user?: GetUserType,
-  },
+  id: ?string,
 };
 
 type State = {
   feed: 'participant' | 'creator',
 };
 
-const UserThreadFeed = compose(getUserThreadConnection)(ThreadFeed);
+const User = compose(getUserById, ViewNetworkHandler)(Profile);
+const CurrentUser = compose(getCurrentUser, ViewNetworkHandler)(Profile);
 
-class User extends React.Component<Props, State> {
-  state = { feed: 'participant' };
-
-  componentDidUpdate() {
-    const { data: { user }, navigation } = this.props;
-    if (!user || navigation.state.params.title) return;
-    navigation.setParams({ title: `${user.name} (@${user.username})` });
-  }
-
-  toggleFeed = feed => this.setState({ feed });
-
+class Container extends React.Component<Props, State> {
   render() {
-    const { data, isLoading, hasError, navigation } = this.props;
-    const { feed } = this.state;
+    const { id, ...rest } = this.props;
 
-    if (data.user) {
-      return (
-        <Wrapper>
-          <StatusBar barStyle="light-content" />
-
-          <ScrollView>
-            <CoverPhotoContainer>
-              <CoverPhoto
-                resizeMode={'cover'}
-                source={{ uri: data.user.coverPhoto }}
-              />
-            </CoverPhotoContainer>
-
-            <ProfilePhotoContainer>
-              <ProfilePhoto source={{ uri: data.user.profilePhoto }} />
-            </ProfilePhotoContainer>
-
-            <ProfileDetailsContainer>
-              <Name>{data.user.name}</Name>
-              <Username>@{data.user.username}</Username>
-              <Description>{data.user.description}</Description>
-            </ProfileDetailsContainer>
-
-            <ThreadFeedTabContainer>
-              <ThreadFeedTab
-                onPress={() => this.toggleFeed('participant')}
-                isActive={feed === 'participant'}
-              >
-                <TabLabel isActive={feed === 'participant'}>Replies</TabLabel>
-              </ThreadFeedTab>
-              <ThreadFeedTab
-                onPress={() => this.toggleFeed('creator')}
-                isActive={feed === 'creator'}
-              >
-                <TabLabel isActive={feed === 'creator'}>Threads</TabLabel>
-              </ThreadFeedTab>
-            </ThreadFeedTabContainer>
-
-            <UserThreadFeed
-              navigation={navigation}
-              kind={this.state.feed}
-              id={data.user.id}
-            />
-          </ScrollView>
-        </Wrapper>
-      );
-    }
-
-    if (isLoading) {
-      return (
-        <Wrapper>
-          <View testID="e2e-User">
-            <Text>Loading...</Text>
-          </View>
-        </Wrapper>
-      );
-    }
-
-    if (hasError) {
-      return (
-        <Wrapper>
-          <View testID="e2e-User">
-            <Text>Error!</Text>
-          </View>
-        </Wrapper>
-      );
-    }
-
-    return null;
+    return id ? <User id={id} {...rest} /> : <CurrentUser {...rest} />;
   }
 }
 
-export default compose(getUserById, ViewNetworkHandler)(User);
+export default Container;
